LayoutTests/platform/chromium/fast/forms/suggestion-picker/date-suggestion-picker-reset-value-after-reload.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/suggestion-picker-common.js"></script>
</head>
<body>
<script>
description('Checks reloading page resets input values');

var iframe;
var testInput1;
var testInput2;

function runOnOpenPopup1() {
    iframe = document.getElementById('iframe');

    popupWindow.focus();

    shouldBeNull('highlightedEntry()');
    eventSender.keyDown('downArrow');
    shouldBe('highlightedEntry()', '"2012-01-01"');
    eventSender.keyDown("\n");
    shouldBeNull('document.getElementById("mock-page-popup")');
    shouldBe('iframe.contentDocument.getElementById("test1").value', '"2012-01-01"');

    openPicker(iframe.contentDocument.getElementById('test2'), runOnOpenPopup2);
}

function runOnOpenPopup2() {
    popupWindow.focus();

    shouldBeNull('highlightedEntry()');
    eventSender.keyDown('downArrow');
    shouldBe('highlightedEntry()', '"2012-01-01"');
    eventSender.keyDown("\n");
    shouldBeNull('document.getElementById("mock-page-popup")');
    shouldBe('iframe.contentDocument.getElementById("test2").value', '"2012-01-01"');

    iframe.addEventListener("load", function () { parent.runOnIFrameLoad(); });
    iframe.contentWindow.location.reload();
}

function runOnIFrameLoad()
{
    testInput1 = iframe.contentDocument.getElementById('test1');
    testInput2 = iframe.contentDocument.getElementById('test2');
    shouldBeEqualToString('testInput1.value', '');
    shouldBeEqualToString('testInput2.value', '2002-02-02');
    iframe.parentNode.removeChild(iframe);
    finishJSTest();
}
</script>
<iframe id=iframe onload="openPicker(event.target.contentDocument.getElementById('test1'), runOnOpenPopup1);" srcdoc="<input type=date id=test1 list=suggestions><input type=date id=test2 value=2002-02-02 list=suggestions><datalist id=suggestions><option>2012-01-01</option></datalist>"></iframe>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
